function/requirements.txt
azure-functions==1.10.1
psycopg2-binary==2.9.3
sendgrid